pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tutorial-teacher"
version = "0.1.0"
description = "Interactive CLI tool for following YouTube coding tutorials with AI assistance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "youtube-transcript-api>=0.6.0",
    "anthropic>=0.18.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
tt = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
